requirements.txt
# FastAPI core
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0

# Database
postgis>=1.0.4,<1.1.0
psycopg2-binary==2.9.3,<2.10.0

# Utilities
requests>=2.27.1,<2.28.0 # for image download
aiofiles>=0.8.0,<0.9.0 # FastAPI dependency for FileResponse
Pillow>=9.0.1,<9.1.0 # for image processing
